+++ spec/integration/api/v2010/account/sip/domain_spec.rb
@@ -53,7 +53,7 @@ describe 'Domain' do
                   "voice_status_callback_method": "POST",
                   "voice_status_callback_url": null,
                   "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                  "trunk_sid": "TRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                  "emergency_calling_enabled": true
               }
           ],
           "start": 0,
@@ -143,7 +143,7 @@ describe 'Domain' do
           "voice_status_callback_method": "POST",
           "voice_status_callback_url": null,
           "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-          "trunk_sid": null
+          "emergency_calling_enabled": true
       }
       ]
     ))
@@ -197,7 +197,7 @@ describe 'Domain' do
           "voice_status_callback_method": "POST",
           "voice_status_callback_url": null,
           "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-          "trunk_sid": "TRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+          "emergency_calling_enabled": true
       }
       ]
     ))
@@ -251,7 +251,7 @@ describe 'Domain' do
           "voice_status_callback_method": "POST",
           "voice_status_callback_url": null,
           "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-          "trunk_sid": "TRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+          "emergency_calling_enabled": true
       }
       ]
     ))