+++ spec/integration/preview/bulk_exports/export/day_spec.rb
@@ -30,7 +30,7 @@ describe 'Day' do
         200,
       %q[
       {
-          "redirect_to": "https://com.twilio.dev-us1.exports.s3.amazonaws.com/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+          "redirect_to": "https://api.bulkexports.twilio.s3.amazonaws.com/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
       }
       ]
     ))